% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa.R
\name{get_lpa_group}
\alias{get_lpa_group}
\title{Get LPA groups}
\usage{
get_lpa_group(G)
}
\arguments{
\item{G}{lpa() output - igraph object with "group" label assigned to each node}
}
\description{
Extract which nodes belong to which LPA cluster. Function requires
graph generated with the lpa() function.
}
