% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa.R
\name{lpa}
\alias{lpa}
\title{Label Propagation Algorithm}
\usage{
lpa(G, method = c("retention", "original"), max_i = 50, verbose = TRUE)
}
\arguments{
\item{G}{igraph undirected graph.}

\item{method}{character value, either "retention" or "original".}

\item{max_i}{integer, maximum number of iterations for algorithm to run.}
}
\value{
igraph object with "group" feature assigned to every node
}
\description{
Run Label Propagation Algorithm on the igraph object. Depending on the method
parameter it will use either retention or original version of the
algorithm. Currently setting initial state is not available and any
value of the "group" label will be overwritten.
}
