% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_lpa.R
\name{visualize_lpa}
\alias{visualize_lpa}
\title{Visualize LPA groups}
\usage{
visualize_lpa(G)
}
\arguments{
\item{G}{lpa() output - igraph object with "group" label assigned to each node}
}
\description{
Creates visNetwork visualization of clusters generated by the LPA.
Function requires graph generated with the lpa() function.
}
